\encoding{UTF-8}

\name{diets-package}

\alias{diets-package}

\alias{diets}

\docType{package}

\title{
\packageTitle{diets}
}

\description{
\packageDescription{diets}
}

\details{

The DESCRIPTION file:
\packageDESCRIPTION{diets}
\packageIndices{diets}
The package is composed of the following parts :

Base: compute average micronutrient intake by person per day, average micronutrient requirements by person per day and deficiencies, from 4 databases : population, micronutrient requirements by age, food consumption (in kCal or grams) and food micronutrient content.

Compute: use the first part to calculate intake and requirements for the globe during the period 1961-2011.

Prepare: use the databases from "Compute" to combine and create indicators.
}

\author{
\packageAuthor{diets}
Maintainer: \packageMaintainer{diets}
}

\references{
IOM (Institute of Medicine) (2011) Dietary Reference Intakes for Calcium and Vitamin D. Washington, DC: The National Academies Press.

Hotz C and Brown KH, (Eds.) (2004) International Zinc Nutrition Consultative Group (IZiNCG) Assessment of the risk of zinc deficiency in populations and   options for its control. Food and Nutrition Bulletin 25: S91-S204, 2004.

FAO & World Health Organization (2004) Vitamin and mineral requirements in human nutrition.

US Department of Agriculture, Agricultural Research Service, Nutrient Data Laboratory (2015) USDA National Nutrient Database for Standard Reference, Release 27 (slightly revised). Version Current:  May 2015.

FAO (2012) West African Food Composition Table. Rome.

University of California at Berkeley International Minilist. WorldFood Dietary Assessment System, 2nd edition.

FAO (Food and Agriculture organisation). Food Balance Sheet (FBS). Documentation link: http://faostat3.fao.org/download/FB/*/E

FAO (Food and Agriculture organisation). Food Balance Sheet (FBS). Download link: http://faostat3.fao.org/faostat-bulkdownloads/FoodBalanceSheets_E_All_Data_(Norm).zip

United Nations (UN) (2012) World Population Database 04.
}

\keyword{ package }
\keyword{ diets }
\keyword{ PIMII }
\keyword{ MDI }
\keyword{ diets }
\keyword{ Intake }
\keyword{ Requirements }
\keyword{ Prevalence }
\keyword{ Deficiencies }
\keyword{ Micronutrients }
\keyword{ Nutrients }
\keyword{ Consumption }
\keyword{ Diet }
\keyword{ Diet quality }
\keyword{ Global Trends }

