\encoding{UTF-8}

\name{diets-package}

\alias{diets-package}

\alias{diets}

\docType{package}

\title{
\packageTitle{diets}
}

\description{
\packageDescription{diets}
}

\details{

This package and data can be used to reproduce and improve estimates of dietary micronutrient supplies and prevalence of inadequate intakes as described by Beal et al https://doi.org/10.1371/journal.pone.0175554
}

\author{
\packageAuthor{diets}
Maintainer: \packageMaintainer{diets}
}

\references{
IOM (Institute of Medicine) (2011) Dietary Reference Intakes for Calcium and Vitamin D. Washington, DC: The National Academies Press.

Hotz C and Brown KH, (Eds.) (2004) International Zinc Nutrition Consultative Group (IZiNCG) Assessment of the risk of zinc deficiency in populations and   options for its control. Food and Nutrition Bulletin 25: S91-S204, 2004.

FAO & World Health Organization (2004) Vitamin and mineral requirements in human nutrition.

US Department of Agriculture, Agricultural Research Service, Nutrient Data Laboratory (2015) USDA National Nutrient Database for Standard Reference, Release 27 (slightly revised). Version Current:  May 2015.

FAO (2012) West African Food Composition Table. Rome.

University of California at Berkeley International Minilist. WorldFood Dietary Assessment System, 2nd edition.

FAO (Food and Agriculture organisation). Food Balance Sheet (FBS). Documentation link: http://faostat3.fao.org/download/FB/*/E

FAO (Food and Agriculture organisation). Food Balance Sheet (FBS). Download link: http://faostat3.fao.org/faostat-bulkdownloads/FoodBalanceSheets_E_All_Data_(Norm).zip

United Nations (UN) (2012) World Population Database 04.
}

\keyword{ package }
